+++ src/Core2D.Core/ViewModels/Shapes/QuadraticBezierShapeViewModel.cs
@@ -94,7 +94,7 @@ public partial class QuadraticBezierShapeViewModel : BaseShapeViewModel
         _point3?.Bind(dataFlow, db, record);
     }
 
-    public override void Move(ISelection? selection, decimal dx, decimal dy)
+    public override void Move(ISelection? selection, double dx, double dy)
     {
         if (_point1 is null || _point2 is null || _point3 is null)
         {
